Correct code size calculations on posts
The "size" of the image file includes the blank space.
Instead, "df" was invokved in the image to get a
count of the size of the used space. There is still
some error here, as there are configuration files, etc.
However, a manual audit of the size shows that most of
it is binary files (libraries or executables).